Change all internal IDs to Longs (not Ints)

+++ app/src/main/java/com/boardgamegeek/db/model/ArtistBriefForUpsert.kt
@@ -4,7 +4,7 @@ import androidx.room.ColumnInfo
 
 data class ArtistBriefForUpsert (
     @ColumnInfo(name = "_id")
-    val internalId: Int,
+    val internalId: Long,
     @ColumnInfo(name = "artist_id")
     val artistId: Int,
     @ColumnInfo(name = "artist_name")
